api/pyproject.toml
[project]
name = "cat-activites-monitor-api"
version = "0.1.0"
description = "A FastAPI app that fetches camera's snapshots and analyzes cats behavior."
authors = [
    {name = "Bianca Rosa", email = "me@biancarosa.com"},
]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "httpx>=0.25.0",
    "pillow>=10.0.0",
    "python-multipart>=0.0.6",
    "apscheduler>=3.10.4",
    "pyyaml>=6.0.1",
    "ultralytics>=8.0.0",
    "opencv-python-headless>=4.8.0",
    "numpy>=1.24.0",
    "scikit-image>=0.21.0",
    "jinja2>=3.1.0",
    "asyncpg>=0.29.0",
]
requires-python = ">=3.11"
license = {text = "MIT"}

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = [
    "main.py",
    "config.yaml",
    "*.md",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
] 
